+++ src/main/java/com/kuit/findyou/global/jwt/argument_resolver/LoginUserIdArgumentResolver.java
@@ -1,5 +1,7 @@
-package com.kuit.findyou.global.common.jwt;
+package com.kuit.findyou.global.jwt.argument_resolver;
 
+import com.kuit.findyou.global.jwt.annotation.LoginUserId;
+import com.kuit.findyou.global.jwt.util.JwtUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
