+++ src/main/java/com/kuit/findyou/global/jwt/util/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.global.common.jwt;
+package com.kuit.findyou.global.jwt.util;
 
 import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Value;
